output: Move file review state assembly
The file-review output layer has dedicated modules for page parsing, layout
sizing, passive model records, and row display ID lookup. The main
file_review module still owns model construction, persisted review-state
assembly, and terminal rendering.

The project lacks a named boundary between rendering and state assembly for
page-aware file reviews. That makes command callers depend on the renderer
for saved review state and keeps footer action selection tied to a large
output module.

This commit addresses that by moving visible action-selection helpers into
output.file_review_action_selections, moving saved-state assembly into
output.file_review_state_builder, and updating show commands to import
state assembly from that module.

Subsequent commits will pin those boundaries in architecture tests and
update command tests to import the state builder directly.
